feat(services): Merge index_manager interfaces and add alias
