+++ src/Resources/config/services.xml
@@ -26,6 +26,8 @@
             <argument type="service" id="algolia_client" />
         </service>
 
+        <service id="Algolia\SearchBundle\IndexManagerInterface" alias="search.index_manager"/>
+
         <!-- CustomNormalizer is not registered by framework-bundle -->
         <service id="custom_normalizer" class="Symfony\Component\Serializer\Normalizer\CustomNormalizer" public="false">
             <tag name="serializer.normalizer" priority="-800" />
